Created AllActivitiesScreen and MapScreen
